+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/req/payload/DaoClaimPayload.java
@@ -2,12 +2,10 @@ package org.nervos.mercury.model.req.payload;
 
 import org.nervos.mercury.model.req.item.Item;
 
-import java.math.BigInteger;
-
 public class DaoClaimPayload {
   public Item from;
   public String to;
-  public BigInteger feeRate;
+  public Long feeRate;
 
   protected DaoClaimPayload() {
   }
