+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/req/since/SinceConfig.java
@@ -2,11 +2,9 @@ package org.nervos.mercury.model.req.since;
 
 import com.google.gson.annotations.SerializedName;
 
-import java.math.BigInteger;
-
 public class SinceConfig {
   public SinceFlag flag;
   @SerializedName("type_")
   public SinceType type;
-  public BigInteger value;
+  public long value;
 }
