+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/resp/GetBalanceResponse.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public class GetBalanceResponse {
   public List<BalanceResponse> balances;
-  public int tipBlockNumber;
+  public long tipBlockNumber;
 }
